+++ src/window.js
@@ -153,7 +153,7 @@ export class Window {
         // remove the element, panes and tabbar gone as they are childs
         this.e.remove()
         // if we're zoomed in, the pane is a child of body
-        if (this.activeP.zoomed)
+        if (this.activeP && this.activeP.zoomed)
             document.body.removeChild(this.activeP.zoomedE)
         this.gate.windows.splice(this.gate.windows.indexOf(this), 1)
         // if we removed a window it means the user can add a window