+++ CHANGELOG.md
@@ -6,7 +6,13 @@ will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.16.0] - 2021/8/15
+## [0.16.1] - 2021/8/15
+
+### Fixed 
+
+- exiting from zoom
+
+## [0.16.0] - 2021/8/11
 
 ### Added